Rename HUD for better component tag auto-detection.
